// lite-api-web/api/api/models/QueryApiInfoPageDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type QueryApiInfoPageDto = {
    current?: number;
    pageSize?: number;
    sortField?: string;
    sortOrder?: string;
    name?: string;
    url?: string;
    status?: number;
    method?: string;
};